extract set_runtime_version and it will always be called
