+++ src/bcbench/evaluate/bugfix.py
@@ -13,6 +13,7 @@ from bcbench.operations import (
     clean_project_paths,
     copy_problem_statement_folder,
     run_tests,
+    set_runtime_version,
     setup_repo_prebuild,
     stage_and_get_diff,
 )
@@ -30,6 +31,7 @@ class BugFixPipeline(EvaluationPipeline[BugFixEntry]):
     def setup_workspace(self, entry: BugFixEntry, repo_path: Path) -> None:
         setup_repo_prebuild(entry, repo_path)
         copy_problem_statement_folder(entry, repo_path)
+        set_runtime_version(entry.project_paths)
 
     def setup(self, context: EvaluationContext[BugFixEntry]) -> None:
         setup_repo_prebuild(context.entry, context.repo_path)
@@ -42,6 +44,7 @@ class BugFixPipeline(EvaluationPipeline[BugFixEntry]):
         )
 
         copy_problem_statement_folder(context.entry, context.repo_path)
+        set_runtime_version(context.entry.project_paths)
 
     def run_agent(self, context: EvaluationContext[BugFixEntry], agent_runner: Callable) -> None:
         with github_log_group(f"{context.agent_name} -- Entry: {context.entry.instance_id}"):
