+++ src/bcbench/operations/__init__.py
@@ -18,7 +18,7 @@ from bcbench.operations.git_operations import (
 from bcbench.operations.hooks_operations import setup_hooks
 from bcbench.operations.instruction_operations import copy_problem_statement_folder, setup_custom_agent, setup_instructions_from_config
 from bcbench.operations.project_operations import categorize_projects
-from bcbench.operations.setup_operations import setup_repo_prebuild
+from bcbench.operations.setup_operations import set_runtime_version, setup_repo_prebuild
 from bcbench.operations.skills_operations import setup_agent_skills
 from bcbench.operations.test_operations import extract_tests_from_patch
 
@@ -36,6 +36,7 @@ __all__ = [
     "copy_problem_statement_folder",
     "extract_tests_from_patch",
     "run_tests",
+    "set_runtime_version",
     "setup_agent_skills",
     "setup_custom_agent",
     "setup_hooks",
