+++ src/bcbench/operations/setup_operations.py
@@ -1,5 +1,6 @@
 """Setup operations for repository preparation."""
 
+import json
 from pathlib import Path
 
 from bcbench.dataset.dataset_entry import BaseDatasetEntry
@@ -9,6 +10,11 @@ from bcbench.operations.git_operations import checkout_commit, clean_repo
 logger = get_logger(__name__)
 
 __all__ = ["setup_repo_prebuild"]
+
+# Offset from BC platform major version to AL runtime version.
+# E.g. platform 25.0 (BC 2024w2) → runtime 14.0, platform 27.0 → runtime 16.0
+# See: BC-DeveloperExperience RuntimeVersion.cs
+_PLATFORM_TO_RUNTIME_OFFSET = 11
 
 
 def setup_repo_prebuild(entry: BaseDatasetEntry, repo_path: Path) -> None:
@@ -28,3 +34,40 @@ def setup_repo_prebuild(entry: BaseDatasetEntry, repo_path: Path) -> None:
 
     clean_repo(repo_path)
     checkout_commit(repo_path, entry.base_commit)
+
+
+def set_runtime_version(project_paths: list[str]) -> None:
+    """Set the AL runtime version in each project's app.json based on platform version.
+
+    The AL compiler (altool) defaults to the latest runtime, enabling newer validation rules that reject older code.
+    Setting the runtime to match the platform version makes the compiler behave like the version that originally compiled the code.
+
+    Can be skippped when altool is not used.
+    """
+    for project_path in project_paths:
+        app_json_path = Path(project_path) / "app.json"
+        if not app_json_path.is_file():
+            continue
+
+        try:
+            app_json = json.loads(app_json_path.read_text(encoding="utf-8-sig"))
+        except (json.JSONDecodeError, OSError):
+            continue
+
+        if app_json.get("runtime"):
+            continue
+
+        platform: str = app_json.get("platform", "")
+        try:
+            platform_major = int(platform.split(".")[0])
+        except (ValueError, IndexError):
+            continue
+
+        runtime_major: int = platform_major - _PLATFORM_TO_RUNTIME_OFFSET
+        if runtime_major < 1:
+            continue
+
+        runtime: str = f"{runtime_major}.0"
+        app_json["runtime"] = runtime
+        app_json_path.write_text(json.dumps(app_json, indent=2, ensure_ascii=False), encoding="utf-8")
+        logger.info(f"Set runtime={runtime} in {app_json_path} (platform {platform})")
