+++ tests/test_mcp_config.py
@@ -5,7 +5,6 @@ from pathlib import Path
 import pytest
 
 from bcbench.agent.shared.altool_paths import build_assembly_probing_paths as _build_assembly_probing_paths
-from bcbench.agent.shared.altool_paths import set_runtime_version as _set_runtime_version
 from bcbench.agent.shared.mcp import build_mcp_config
 from tests.conftest import create_dataset_entry
 
@@ -194,35 +193,3 @@ class TestBuildAssemblyProbingPaths:
         result = _build_assembly_probing_paths(tmp_path)
 
         assert isinstance(result, list)
-
-
-class TestSetRuntimeVersion:
-    def test_sets_runtime_from_platform(self, tmp_path):
-        app_json = {"platform": "25.0.0.0", "version": "25.0.0.0"}
-        (tmp_path / "app.json").write_text(json.dumps(app_json))
-
-        _set_runtime_version([str(tmp_path)])
-
-        result = json.loads((tmp_path / "app.json").read_text())
-        assert result["runtime"] == "14.0"
-
-    def test_skips_when_runtime_already_set(self, tmp_path):
-        app_json = {"platform": "25.0.0.0", "runtime": "12.0"}
-        (tmp_path / "app.json").write_text(json.dumps(app_json))
-
-        _set_runtime_version([str(tmp_path)])
-
-        result = json.loads((tmp_path / "app.json").read_text())
-        assert result["runtime"] == "12.0"
-
-    def test_platform_27_maps_to_runtime_16(self, tmp_path):
-        app_json = {"platform": "27.0.0.0"}
-        (tmp_path / "app.json").write_text(json.dumps(app_json))
-
-        _set_runtime_version([str(tmp_path)])
-
-        result = json.loads((tmp_path / "app.json").read_text())
-        assert result["runtime"] == "16.0"
-
-    def test_skips_missing_app_json(self, tmp_path):
-        _set_runtime_version([str(tmp_path)])  # should not raise
